# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: DEVLINK API
  description: API service for the Distributed Link Intelligence Platform.
  version: 0.1.0
paths:
...
  /r/{short_code}:
    get:
      tags:
      - redirect
      summary: Resolve a short code
      description: Public redirect hot path. Looks up link metadata from Redis first,
        falls back to PostgreSQL, records a non-blocking click event, and returns
        a temporary redirect when the link is active.
      operationId: redirect_short_code_r__short_code__get
      parameters:
      - name: short_code
        in: path
        required: true
        schema:
          type: string
          title: Short Code
      - name: redirect_token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Redirect Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '401':
          description: Password-protected redirect token is missing or invalid.
        '403':
          description: Link is blocked for safety reasons.
        '404':
          description: Short code not found.
        '410':
          description: Link is inactive or expired.
        '422':
          description: Request validation failed.
        '429':
          description: Redirect or password verification rate limit exceeded.
  /r/{short_code}/verify:
    post:
      tags:
      - redirect
      summary: Verify a protected redirect
      description: Verify a password for a protected short link and return a short-lived
        redirect token. Attempts are rate-limited.
      operationId: verify_redirect_password_r__short_code__verify_post
      parameters:
      - name: short_code
        in: path
        required: true
        schema:
          type: string
          title: Short Code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordVerifyRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordVerifyResponse'
        '401':
          description: Password-protected redirect token is missing or invalid.
        '403':
          description: Link is blocked for safety reasons.
        '404':
          description: Short code not found.
        '410':
          description: Link is inactive or expired.
        '422':
          description: Request validation failed.
        '429':
          description: Redirect or password verification rate limit exceeded.
...
components:
  schemas:
...
    PasswordVerifyRequest:
      properties:
        password:
          type: string
          maxLength: 128
          minLength: 1
          title: Password
      additionalProperties: false
      type: object
      required:
      - password
      title: PasswordVerifyRequest
      description: Request body for password-protected redirect verification.
      example:
        password: optional-secret
    PasswordVerifyResponse:
      properties:
        redirect_token:
          type: string
          title: Redirect Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
      type: object
      required:
      - redirect_token
      - expires_in
      title: PasswordVerifyResponse
      description: Response payload for successful redirect password verification.
      example:
        expires_in: 600
        redirect_token: eyJhbGciOiJIUzI1NiJ9...
        token_type: bearer
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login